requirements.txt
Flask
Flask-SocketIO
llama-index-core
llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-huggingface